Setup Github Actions
Basic workflow for Github Actions, building only the core files. It includes some minor fixes so it builds in all three platforms.

+++ core/src/Simulation/Simulation.cpp
@@ -189,8 +189,8 @@ bool Simulation::update( void ) noexcept
     auto frameEndTime = clock::now();
     auto delta = frameEndTime - frameStartTime;
     auto t = std::max(
-        1ll,
-        ( MIN_FRAME_TIME - std::chrono::duration_cast< std::chrono::nanoseconds >( delta ) ).count() );
+        Int64( 1 ),
+        Int64( ( MIN_FRAME_TIME - std::chrono::duration_cast< std::chrono::nanoseconds >( delta ) ).count() ) ); 
     std::this_thread::sleep_for( std::chrono::nanoseconds( t ) );
 #endif
